Provided error detection if resources are missing
If the following resources: 7za, exclusionlist, and or exclusiondirlist
was not found, this program will notify the user that the resources were
not located and that the user should report the issue.  However, in
addition, this program will automatically open the default web-broswer
to the issues page in the WolfenDoom project via GitHub web services.