Provided custom window titles to the program
This program will now have a unique title when it starts up, but during
close - it will revert to the traditional title.  It is possible to
fetch the current title of the already working shell before the program
starts, but the work is not worth the effort for such a small feature.