Improved git feature check handling
Refined the checking to also include finding the .git directory within
the source code.  This is done to prevent issues when featuresGit is
true - because the git.exe was detected but if .git directory does not
exist (for whatever reason), then an error will pop up within the
buffer.  To compact this issue, except only checking for the git.exe,
also check for the .git directory within the source code before setting
the featuresGit to true.  If either of these detections fails, then the
feature is ultimately false.  Thus, if false - then the feature within
the program is disabled.